first auth usage
