+++ packages/common/src/index.ts
@@ -1,7 +1,10 @@
 import moment from 'moment';
+import { LightdashUser, OrganizationUser } from './types/user';
 
 export * from './authorization/organizationMemberAbility';
+export * from './types/organization';
 export * from './types/organizationMemberProfile';
+export * from './types/user';
 
 const DATE_FORMAT = 'YYYY-MM-DD';
 export const formatDate = (date: Date): string =>
@@ -579,55 +582,6 @@ export type ApiExploreResults = Explore;
 export type ApiStatusResults = 'loading' | 'ready' | 'error';
 
 export type ApiRefreshResults = undefined;
-
-export interface LightdashUser {
-    userUuid: string;
-    email: string | undefined;
-    firstName: string;
-    lastName: string;
-    organizationUuid: string;
-    organizationName: string;
-    isTrackingAnonymized: boolean;
-    isMarketingOptedIn: boolean;
-    isSetupComplete: boolean;
-}
-
-export interface SessionUser extends LightdashUser {
-    userId: number;
-}
-
-export const isSessionUser = (user: any): user is SessionUser =>
-    typeof user === 'object' &&
-    user !== null &&
-    user.userUuid &&
-    user.userId &&
-    user.openId === undefined;
-
-export interface OpenIdUser {
-    openId: {
-        subject: string;
-        issuer: string;
-        email: string;
-        firstName: string | undefined;
-        lastName: string | undefined;
-    };
-}
-
-export const isOpenIdUser = (user: any): user is OpenIdUser =>
-    typeof user === 'object' &&
-    user !== null &&
-    user.userUuid === undefined &&
-    user.userId === undefined &&
-    typeof user.openId === 'object' &&
-    user.openId !== null &&
-    typeof user.openId.subject === 'string' &&
-    typeof user.openId.issuer === 'string' &&
-    typeof user.openId.email === 'string';
-
-export type OrganizationUser = Pick<
-    LightdashUser,
-    'userUuid' | 'firstName' | 'lastName' | 'email'
->;
 
 export type CreateUserArgs = {
     firstName: string;
