+++ packages/common/src/types/organization.ts
@@ -1,0 +1,3 @@
+export type Organization = {
+    organizationUuid: string;
+};
